+++ test/data/category.js
@@ -1,0 +1,24 @@
+const categories = [
+    {
+        id: 1,
+        nameKr: 'OTT',
+        nameEng: 'movie',
+        description: '영화/드라마',
+        smallLogoPath: '/images/category/icon-fill-20-book@2x.png',
+        bigLogoPath: '/images/category/list-icon-category-book@2x.png',
+        sort: 1,
+        template: '1 2 2 2 1',
+    },
+    {
+        id: 2,
+        nameKr: '음악',
+        nameEng: 'music',
+        description: '음악',
+        smallLogoPath: '/images/category/icon-fill-20-music@2x.png',
+        bigLogoPath: '/images/category/list-icon-category-music@2x.png',
+        sort: 2,
+        template: '1 1 1 1 1',
+    },
+];
+
+module.exports = categories;
