+++ test/data/comparisonItem.js
@@ -1,0 +1,84 @@
+const comparisonItems = [
+    {
+        id: 1,
+        code: 'MONTH_FEE',
+        name: '월 구독료',
+        unit: '원',
+        type: 'BARCHART',
+        imgPath: '/images/comparisonItem/pay.png',
+        sort: 1,
+        categoryId: 1,
+    },
+    {
+        id: 2,
+        code: 'SERVICE',
+        name: '제공서비스',
+        unit: '',
+        type: 'TEXT',
+        imgPath: '/images/comparisonItem/rocket.png',
+        sort: 2,
+        categoryId: 1,
+    },
+    {
+        id: 3,
+        code: 'ACCESS_COUNT',
+        name: '동시 접속 가능 인원',
+        unit: '명',
+        type: 'TEXT',
+        imgPath: '/images/comparisonItem/hand.png',
+        sort: 3,
+        categoryId: 1,
+    },
+    {
+        id: 4,
+        code: 'PAYMENT_UNIT',
+        name: '결제단위',
+        unit: '',
+        type: 'TEXT',
+        imgPath: '/images/comparisonItem/pay.png',
+        sort: 4,
+        categoryId: 1,
+    },
+    {
+        id: 5,
+        code: 'USING_RANGE',
+        name: '이용범위',
+        unit: '',
+        type: 'TEXT',
+        imgPath: '/images/comparisonItem/glass-tilted.png',
+        sort: 5,
+        categoryId: 1,
+    },
+    {
+        id: 6,
+        code: 'FREE_PERIOD',
+        name: '무료기간',
+        unit: '',
+        type: 'TEXT',
+        imgPath: '/images/comparisonItem/free.png',
+        sort: 6,
+        categoryId: 1,
+    },
+    {
+        id: 7,
+        code: 'ADDITIONAL',
+        name: '부가기능 및 기타장 강점',
+        unit: '',
+        type: 'TEXT',
+        imgPath: '/images/comparisonItem/wrench.png',
+        sort: 7,
+        categoryId: 1,
+    },
+    {
+        id: 8,
+        code: 'CONTENTS_COUNT',
+        name: '컨텐츠 수',
+        unit: '',
+        type: 'DOUGHNUTCHART',
+        imgPath: '/images/comparisonItem/videocassette.png',
+        sort: 8,
+        categoryId: 1,
+    },
+];
+
+module.exports = comparisonItems;
