+++ test/data/comparisonValue.js
@@ -1,0 +1,201 @@
+const comparisonValues = [
+    // 넷플릭스 베이식
+    {
+        id: 1,
+        value: '9500',
+        comparisonItemId: 1,
+        membershipId: 1,
+    },
+    {
+        id: 2,
+        value: 'TV 프로그램과 영화를 광고 없이 시청할 수 있는 스트리밍 서비스',
+        comparisonItemId: 2,
+        membershipId: 1,
+    },
+    {
+        id: 3,
+        value: '1',
+        comparisonItemId: 3,
+        membershipId: 1,
+    },
+    {
+        id: 4,
+        value: '한 달 단위',
+        comparisonItemId: 4,
+        membershipId: 1,
+    },
+    {
+        id: 5,
+        value: '모든 인터넷 연결 지원 디바이스',
+        comparisonItemId: 5,
+        membershipId: 1,
+    },
+    {
+        id: 6,
+        value: '신규회원 한정 1개월 무료',
+        comparisonItemId: 6,
+        membershipId: 1,
+    },
+    {
+        id: 7,
+        value: '백그라운드 재생\nKT멤버쉽 포인트 사용 가능',
+        comparisonItemId: 7,
+        membershipId: 1,
+    },
+    {
+        id: 8,
+        value: '{ "labels": ["영화", "기타"], "data": ["20000", "50000"] }',
+        comparisonItemId: 8,
+        membershipId: 1,
+    },
+    // 넷플릭스 스텐다드
+    {
+        id: 9,
+        value: '12000',
+        comparisonItemId: 1,
+        membershipId: 2,
+    },
+    {
+        id: 10,
+        value: 'TV 프로그램과 영화를 광고 없이 시청할 수 있는 스트리밍 서비스',
+        comparisonItemId: 2,
+        membershipId: 2,
+    },
+    {
+        id: 11,
+        value: '2',
+        comparisonItemId: 3,
+        membershipId: 2,
+    },
+    {
+        id: 12,
+        value: '한 달 단위',
+        comparisonItemId: 4,
+        membershipId: 2,
+    },
+    {
+        id: 13,
+        value: '모든 인터넷 연결 지원 디바이스',
+        comparisonItemId: 5,
+        membershipId: 2,
+    },
+    {
+        id: 14,
+        value: '신규회원 한정 1개월 무료',
+        comparisonItemId: 6,
+        membershipId: 2,
+    },
+    {
+        id: 15,
+        value: 'KT멤버쉽 포인트 사용 가능\n넷플릭스 오리지널 제공',
+        comparisonItemId: 7,
+        membershipId: 2,
+    },
+    {
+        id: 16,
+        value: '{ "labels": ["영화", "기타"], "data": ["20000", "50000"] }',
+        comparisonItemId: 8,
+        membershipId: 2,
+    },
+    // 넷플릭스 프리미엄
+    {
+        id: 17,
+        value: '17000',
+        comparisonItemId: 1,
+        membershipId: 3,
+    },
+    {
+        id: 18,
+        value: 'TV 프로그램과 영화를 광고 없이 시청할 수 있는 스트리밍 서비스',
+        comparisonItemId: 2,
+        membershipId: 3,
+    },
+    {
+        id: 19,
+        value: '4',
+        comparisonItemId: 3,
+        membershipId: 3,
+    },
+    {
+        id: 20,
+        value: '한 달 단위',
+        comparisonItemId: 4,
+        membershipId: 3,
+    },
+    {
+        id: 21,
+        value: '모든 인터넷 연결 지원 디바이스',
+        comparisonItemId: 5,
+        membershipId: 3,
+    },
+    {
+        id: 22,
+        value: '신규회원 한정 1개월 무료',
+        comparisonItemId: 6,
+        membershipId: 3,
+    },
+    {
+        id: 23,
+        value: 'KT멤버쉽 포인트 사용 가능\n넷플릭스 오리지널 제공',
+        comparisonItemId: 7,
+        membershipId: 3,
+    },
+    {
+        id: 24,
+        value: '{ "labels": ["영화", "기타"], "data": ["20000", "50000"] }',
+        comparisonItemId: 8,
+        membershipId: 3,
+    },
+
+    // 유튜브 프리미엄
+    {
+        id: 25,
+        value: '10450',
+        comparisonItemId: 1,
+        membershipId: 4,
+    },
+    {
+        id: 26,
+        value: '다양한 컨텐츠를 광고 없이 시청할 수 있는 서비스',
+        comparisonItemId: 2,
+        membershipId: 4,
+    },
+    {
+        id: 27,
+        value: '1',
+        comparisonItemId: 3,
+        membershipId: 4,
+    },
+    {
+        id: 28,
+        value: '한 달 단위',
+        comparisonItemId: 4,
+        membershipId: 4,
+    },
+    {
+        id: 29,
+        value: '모든 인터넷 연결 지원 디바이스',
+        comparisonItemId: 5,
+        membershipId: 4,
+    },
+    {
+        id: 30,
+        value: '신규회원 한정 1개월 무료',
+        comparisonItemId: 6,
+        membershipId: 4,
+    },
+    {
+        id: 31,
+        value: '광고제거, 영상 다운로드, 백그라운드 재생\nSK, LG U+, KT 통신사 제휴',
+        comparisonItemId: 7,
+        membershipId: 4,
+    },
+    {
+        id: 32,
+        value: '{ "labels": ["영화", "기타"], "data": ["20000", "50000"] }',
+        comparisonItemId: 8,
+        membershipId: 4,
+    },
+];
+
+module.exports = comparisonValues;
