+++ test/data/hashtag.js
@@ -1,0 +1,19 @@
+const hashtag = [
+    {
+        id: 1,
+        name: '영화',
+        categoryId: 1,
+    },
+    {
+        id: 2,
+        name: '드라마',
+        categoryId: 1,
+    },
+    {
+        id: 3,
+        name: '예능',
+        categoryId: 1,
+    },
+];
+
+module.exports = hashtag;
