+++ test/data/membership.js
@@ -1,0 +1,28 @@
+const memberships = [
+    {
+        id: 1,
+        name: '베이식',
+        grade: 1,
+        subscriptionServiceId: 1,
+    },
+    {
+        id: 2,
+        name: '스탠다드',
+        grade: 2,
+        subscriptionServiceId: 1,
+    },
+    {
+        id: 3,
+        name: '프리미엄',
+        grade: 3,
+        subscriptionServiceId: 1,
+    },
+    {
+        id: 4,
+        name: '기본',
+        grade: 1,
+        subscriptionServiceId: 2,
+    },
+];
+
+module.exports = memberships;
