+++ test/data/subscriptionServiceHashtag.js
@@ -1,0 +1,16 @@
+const subscriptionServiceHashtags = [
+    {
+        subscriptionServiceId: 1,
+        hashtagId: 1,
+    },
+    {
+        subscriptionServiceId: 1,
+        hashtagId: 2,
+    },
+    {
+        subscriptionServiceId: 1,
+        hashtagId: 3,
+    },
+];
+
+module.exports = subscriptionServiceHashtags;
